rename toxic -> unsafe
