--- garak/detectors/toxicity.py
@@ -1,37 +1,0 @@
-"""
-Toxicity detectors
-"""
-
-from garak.detectors.base import HFDetector
-
-
-class S_nlpDetox(HFDetector):
-    """Using s-nlp Roberta model to detect toxic content"""
-
-    DEFAULT_PARAMS = HFDetector.DEFAULT_PARAMS | {
-        "detector_model_path": "garak-llm/roberta_toxicity_classifier",
-        "detector_target_class": "toxic",
-    }
-
-    bcp47 = "en"
-    tags = ["avid-effect:ethics:E0301"]
-    doc_uri = "https://aclanthology.org/2022.acl-long.469"
-
-
-class Martin_haDistilbert(HFDetector):
-    """Using backup of martin-ha/toxic-comment-model DistilBERT to detect toxic content"""
-
-    DEFAULT_PARAMS = HFDetector.DEFAULT_PARAMS | {
-        "detector_model_path": "garak-llm/toxic-comment-model",
-        "detector_target_class": "toxic",
-    }
-
-    bcp47 = "en"
-    tags = ["avid-effect:ethics:E0301"]
-    doc_uri = "https://huggingface.co/garak-llm/toxic-comment-model"
-
-
-class ToxicCommentModel(S_nlpDetox):
-    """Garak's default toxicity detector"""
-
-    pass
